Execute order before activity
This is a test to check if there is a way for units to goto and finish an activity in a single turn. Currently seems to work for sequential stuff, like "connect with roads" commands. But when
given a simple "goto and do something" the unit stays still for a turn and does both in the next turn. Thereby still taking 2 turns.